refactor the locale parser
