+++ src/layout/GlobalProvider/index.tsx
@@ -1,20 +1,18 @@
 import { cookies, headers } from 'next/headers';
 import { PropsWithChildren, Suspense } from 'react';
-import { resolveAcceptLanguage } from 'resolve-accept-language';
 
 import { appEnv } from '@/config/app';
 import { getServerFeatureFlagsValue } from '@/config/featureFlags';
-import { DEFAULT_LANG, LOBE_LOCALE_COOKIE } from '@/const/locale';
+import { LOBE_LOCALE_COOKIE } from '@/const/locale';
 import {
   LOBE_THEME_APPEARANCE,
   LOBE_THEME_NEUTRAL_COLOR,
   LOBE_THEME_PRIMARY_COLOR,
 } from '@/const/theme';
 import DebugUI from '@/features/DebugUI';
-import { locales } from '@/locales/resources';
 import { getServerGlobalConfig } from '@/server/globalConfig';
 import { ServerConfigStoreProvider } from '@/store/serverConfig';
-import { getAntdLocale } from '@/utils/locale';
+import { getAntdLocale, parseBrowserLanguage } from '@/utils/locale';
 import { isMobileDevice } from '@/utils/server/responsive';
 
 import AntdV5MonkeyPatch from './AntdV5MonkeyPatch';
@@ -25,30 +23,6 @@ import ReactScan from './ReactScan';
 import StoreInitialization from './StoreInitialization';
 import StyleRegistry from './StyleRegistry';
 
-const parserFallbackLang = async () => {
-  // if the default language is not 'en-US', just return the default language as fallback lang
-  if (DEFAULT_LANG !== 'en-US') return DEFAULT_LANG;
-
-  const header = await headers();
-  /**
-   * The arguments are as follows:
-   *
-   * 1) The HTTP accept-language header.
-   * 2) The available locales (they must contain the default locale).
-   * 3) The default locale.
-   */
-  let fallbackLang: string = resolveAcceptLanguage(
-    header.get('accept-language') || '',
-    //  Invalid locale identifier 'ar'. A valid locale should follow the BCP 47 'language-country' format.
-    locales.map((locale) => (locale === 'ar' ? 'ar-EG' : locale)),
-    DEFAULT_LANG,
-  );
-  // if match the ar-EG then fallback to ar
-  if (fallbackLang === 'ar-EG') fallbackLang = 'ar';
-
-  return fallbackLang;
-};
-
 const GlobalLayout = async ({ children }: PropsWithChildren) => {
   // get default theme config to use with ssr
   const cookieStore = await cookies();
@@ -58,7 +32,8 @@ const GlobalLayout = async ({ children }: PropsWithChildren) => {
 
   // get default locale config to use with ssr
   const defaultLang = cookieStore.get(LOBE_LOCALE_COOKIE);
-  const fallbackLang = await parserFallbackLang();
+  const header = await headers();
+  const fallbackLang = parseBrowserLanguage(header);
 
   // if it's a new user, there's no cookie
   // So we need to use the fallback language parsed by accept-language
