+++ src/utils/locale.ts
@@ -1,4 +1,7 @@
-import { normalizeLocale } from '@/locales/resources';
+import { resolveAcceptLanguage } from 'resolve-accept-language';
+
+import { DEFAULT_LANG } from '@/const/locale';
+import { locales, normalizeLocale } from '@/locales/resources';
 
 export const getAntdLocale = async (lang?: string) => {
   let normalLang = normalizeLocale(lang);
@@ -13,4 +16,31 @@ export const getAntdLocale = async (lang?: string) => {
   const { default: locale } = await import(`antd/locale/${normalLang.replace('-', '_')}.js`);
 
   return locale;
+};
+
+/**
+ * Parse the browser language and return the fallback language
+ * @param headers
+ */
+export const parseBrowserLanguage = (headers: Headers) => {
+  // if the default language is not 'en-US', just return the default language as fallback lang
+  if (DEFAULT_LANG !== 'en-US') return DEFAULT_LANG;
+
+  /**
+   * The arguments are as follows:
+   *
+   * 1) The HTTP accept-language header.
+   * 2) The available locales (they must contain the default locale).
+   * 3) The default locale.
+   */
+  let fallbackLang: string = resolveAcceptLanguage(
+    headers.get('accept-language') || '',
+    //  Invalid locale identifier 'ar'. A valid locale should follow the BCP 47 'language-country' format.
+    locales.map((locale) => (locale === 'ar' ? 'ar-EG' : locale)),
+    DEFAULT_LANG,
+  );
+  // if match the ar-EG then fallback to ar
+  if (fallbackLang === 'ar-EG') fallbackLang = 'ar';
+
+  return fallbackLang;
 };
